Reorganized directory to be a module (for zipapp)
