+++ src/__main__.py
@@ -1,0 +1,3 @@
+import builder.launchgui
+
+builder.launchgui.main()
