+++ src/builder/devtesting.py
@@ -6,9 +6,9 @@ development.
 '''
 
 from typing import List
-import proc_constraints as proc
-import io_file
-import models
+import builder.proc_constraints as proc
+import builder.io_file as io_file
+import builder.models as models
 
 
 # def dummyOldProjectState() -> models.OLDProjectState:
