+++ src/builder/gui_newproject.py
@@ -9,14 +9,14 @@ from pathlib import Path
 from tkinter import ttk
 from typing import List
 
-import proc_constraints as proc
-import gui_projectoverview
-import io_file
-import linting as lint
-import models
-from gui_consts import *
-import devtesting
-import proc_render as render
+import builder.proc_constraints as proc
+import builder.gui_projectoverview as gui_projectoverview
+import builder.io_file as io_file
+import builder.linting as lint
+import builder.models as models
+import builder.devtesting as devtesting
+import builder.proc_render as render
+from builder.gui_consts import *
 
 # Exposed gui elements
 _lblObjFile: tk.Label = None
